[JIT] Propagate requires_grad to autodiff subgraphs
When JIT autodiff is constructing a gradient computation graph, it will only add gradients for tensors that require_grad. Previously, require_grad information was **not** propagated to the subgraph that autodiff used; as a result, autodiff would calculate *all* gradients, even if requires_grad had never been set during profiling runs. In certain cases, this can lead to performance issues. For example, during training, the gradient of the input data is not needed, but is still computed.

This propagates requires_grad to the subgraph passed into autodiff, so that autodiff will not compute unnecessary gradients.

Test: `./bin/test_jit --gtest_filter="JitProfilingFixtureTest.RemoveUnusedLinearGradientComputations"

ghstack-source-id: 1d320b2b8f9356c7900650987608f7d663dded5e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/71666